ruby plugin: be smarter about arch-specific paths (#1730)
Ruby uses some pretty convoluted rules for generating its arch-specific
paths, some of which is hard-coded. This path follows a different
pattern depending on the architecture on which it runs. As a result, the
current Ruby plugin doesn't work on any arch other than amd64 and i386.

Rather than try and be smart about these paths, use globbing to detect
them using files we know are in there. That will allow this plugin to
start working across all supported architectures.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>