tests: add the home plug to the plainbox snap
On integration tests we set XDG_CACHE_HOME to a temporary directory in TMPDIR.
In autopkgtests we are now setting TMPDIR to $HOME/autopkgtest_tmp.
The plainbox test tries to write to XDG_CACHE_HOME, which used to be somewhere
it could write. Now that it is in home, to properly work the snap requires the
home plug.

LP: #1732076