ament plugin: new plugin (#1583)
As ROS2 nears a stable release, it's a good time to introduce support
for it by way of a beta plugin. Start simple, just bootstrap ROS2 and
use it to build the provided source. ROS2 doesn't have a lot of tooling
yet, so this plugin will need to grow with it.

LP: #1686850
Fixes #1443

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>