snapcraft.yaml: use gcc to determine tuple (#1764)
Currently, when building the Snapcraft snap, we try to create the
libsodium.so symlink by manually constructing the arch tuple using
`uname`. However, this does not produce the correct tuple in all cases,
particularly i686.

Use `gcc -print-multiarch` instead. It's exactly what we need, and it's
already a `build-package`.

LP: #1733922

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>