ci: name the travis jobs
It is really hard to have a general view of what is failing
in the eagle eye view offered by travis when using stages and
parallel jobs inside.

We introduce use of the TEST travis defined environment variable
as a mechanism to view what is being run in each job.

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>