many: set rpath for elf files for classic
When working with classic confinement `LD_LIBRARY_PATH`
cannot be set without consequences
on when other assets on the host are
called from the snap.

This sets an rpath for every elf file
when confinement is set to classic
and allows for further extending this
mechanism to non classic.

When setting rpaths it also makes
use of $ORIGIN to setup relative paths.

Fixes: #1663
Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>