project_loader: support grammar on architectures
Currently Snapcraft supports a flat list of architecture overrides in
the snapcraft.yaml that look something like this:

    architectures: [amd64, i386]

If this property is specified, Snapcraft passes it on verbatim to the
snap.yaml, which then assumes the same generated snap supports both of
those architectures (called a "fat" snap).

Expand Snapcraft's `architectures` property to support the advanced
grammar, which allows the property to be dual-purposed into not only
run-time architecture support, but build-time (i.e. specify the
architecture on which it should be built, as well as the architectures
on which the resulting snap can be run). That looks something like this:

    architectures:
      - on i386: [amd64, i386]
      - on armhf: [arm64, armhf]

The final snap.yaml will still only contain a flat list of
architectures.

Resolve #1681

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>