tests: move the python plugin integration tests to a separate suite
The plugins suite is again causing time outs because it's too big.
We discussed about an ugly hack to split the suite, and we already applied it to the catkin plugin. Here we are doing the same for python, and we might have to do it again for ruby or catkin.