go plugin: strip sections that patchelf does not handle
patchelf does not handle the existence of the .note.go.buildid section
so we strip it from the resulting binary as from the perspective of the
snap it is not needed.

A newer patchelf is also required which handles the math required for
the way the headers are layed out when using golang's go.

LP: #1736861

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>