zip: support extracting non-unix zip files
Currently, Snapcraft assumes that the external_attr attribute of each
item in the ZipFile contains unix permissions. That unfortunately isn't
the case of the zip file was created on a non-unix system, and such zip
files cause Snapcraft to create files and directories with 0o000
permissions.

Check the extracted mode before applying it, and don't use chmod if the
mode is 0o000.

LP: #1742373

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>