elf: cleaner patchelf experience
Use the patchelf on the system by default if not running from the
snap and if it fails instruct the user to use patchelf as a part
or as a build-snaps entry.

Given that when using as a part the user can choose to not use
filters to remove the file from prime, we also make a copy of
the file to patch (breaking any potential existing hard link
with the other steps) to avoid the problem of trying to patch
the patchelf primed file with the same patchelf binary file.

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>