elf: make patchelf a dependency
Making it a build_tools entry in the project_loader was the wrong
call, we do not for example do the same for xdelta3. We should
instead be handling it through the delivery mechanisms in place,
such that it is in debian/control for the deb and bundled in the
snapcraft snap when running from there.

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>