cli: use C.UTF-8 if locale not set
When running Snapcraft in a specific way (e.g. Docker, LXD) it's
possible to have a situation where no locale is set. This happens mostly
in CI systems. Click raises a `RuntimeError` in this situation, which is
a pretty terrible user experience.

If C.UTF-8 is available, default to using it. This will allow Snapcraft
to run on these systems without creating more work for our users.

LP: #1746100

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>