storeapi: handle errors even for >400 responses
The requests lib's Response object overwrites __bool__ to return false
if the response code indicates a failure. As a result, if we want to see
if a response is actually `None` as must compare directly to `None`. Or
just make a call and catch a missing attribute which is what we do here.

LP: #1736963
LP: #1748802
LP: #1699232

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>