elf: better debug messages
Some debug messages needed polishing, like the linker check.
Others, like the check for exclusions of libraries, needed
polish to not repeat itself over and over again when
running off of a non supported base. This comes with a
side effect of reducing the amount of calls necessary
for the check where the debug message is displayed.

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>