catkin plugin: support recursive rosinstall files (#1934)
Currently, the Catkin plugin supports rosinstall files in that, if you
hand it a list of rosinstall files in the source, it'll merge them into
the workspace and fetch the repositories listed therein. However, what
happens if _those_ repositories have rosinstall files of their own? This
is not a situation wstool handles, and is something vendors script
around.

Add explicit support in Snapcraft for this situation by way of a new
boolean option: `recursive-rosinstall`. This option defaults to false
(since that's typically how wstool behaves), but if it's true, it'll
continue merging/updating the workspace with newly-fetched rosinstall
files until no more are found.

Closes: #1714

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>