elf: only patch elf files that aren't referenced by DT_NEEDED (#1959)
In brief, when the dynamic linker is searching for DT_NEEDED dependencies of a library, it can use the rpath of the executable or library that pulled in the library. So for classic snaps, it should only be necessary to patch the rpath of executables or shared library plugins. As a simple heuristic, any elf file whose soname is listed in the DT_NEEDED of some other elf file doesn't need patching.

This also updates ElfFile to extract architecture information from elf files. This way we don't e.g. treat an ARM library as satisfying a dependency of an AMD64 executable.

LP: #1747097