pluginhandler: special case go patchelf failures for classic confinement
Classic confinement for go with patchelf is still a little problematic
and has never really work from a confinement perspective. So we warn if
patchelf fails to run through.

Clean up the logic after the introduction of bases and the requirement
to explicitly add libc6 to stage-packages in case of non host
compatibilities