many: add snapcraftctl command for scriptlets (#2002)
There is the desire for scriptlets to be able to communicate with
Snapcraft, e.g. to request that it use a specific name, or version, or
do a specific action, such as fetch stage packages, or run a lifecycle
step.

Add the basic framework for this communication by adding a new CLI
utility called `snapcraftctl` that is used to communicate with the
parent Snapcraft process. Add a new subcommand to this utility called
`build` which, if called, will run the plugin's build step. Also add the
`overide-build` scriptlet, which can use `snapcraftctl build` like so:

    # ...
    override-build: |
      snapcraftctl build
    # ...

Resolve #1675
LP: #1751749

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>