meta: implement passthrough of properties to snap.yaml (#2053)
Introduce a new `passthrough` property, both at the root and in `apps` as
well as `hooks`. This property is defined as a dictionary (object in the
schema) and not validated, but instead applied when generating the 
`snap.yaml`. Its use is logged and ambiguous keys will raise an error. 
The original YAML is used to validate what values have been specified as 
opposed to default values filled in by Snapcraft.

Resolve #2026