python: bring back support for older versions of pip (#2055)
Bring back support for older versions of pip, most notably 8.1.1 which is in 16.04, that don't
support `pip list --format=json`. Ease of parsing is why we dropped that support previously.
And we support pulling in pip from `stage-packages` so that should include 16.04 and not
require resorting to PyPI, which is what snapcraft does with `pip`, `setuptools` and `wheel`
when they're not available.