meta: soften warning about using passthrough (#2091)
Currently Snapcraft warns that a snap using passthrough cannot be
released to the store. What it means is that it may be flagged for
manual review upon push, but even that isn't necessarily the case, as
the store may gain support for the option being passed through. Also,
some users are interpreting this warning as a fatal error, i.e. that
Snapcraft won't allow pushes (or a release).

Stop mentioning the store in the passthrough warning.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>