tests: adapt the integration suite to work for all releases
Setup libc6 as a stage-packages entry were necessary, or alternatively
set the base to core18 if it will not leat to a scenario where it
needs to be downloaded.

Also remove a test which cannot be packaged in bionic as a deb source
anymore, it is unfortunate but it is also a test which may not be
necessary to test as an integration test anymore.

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>