repo: automatically prune unneeded stage-packages
The apt cache isn't too bright when it comes to alternative
dependencies. As a result, it's possible for the snapcraft CLI to get
into the following situation:

- A depends on B OR C
- B depends on D
- C conflicts with B
- A and C are `stage-packages`

If A is handled first, apt defaults to the first alternate (which is
sensible), and marks both A and B (and, by extension, D) to be
installed. Then C is handled, which automatically unmarks B, but doesn't
do anything to D. As a result, A, C, and D are installed into the snap,
even though D is no longer required.

Fix this by crawling the set of changes in the cache and taking care of
anything that can be automatically removed.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>