cli: add provides command
Once the snapcraft CLI merges files from multiple parts into the staging
and priming areas, it's easy to lose track of which part provided which
file. Add a new `snapcraft provides` command meant to serve exactly that
purpose.

    $ snapcraft provides <path>

`<path>` must be in either the staging or priming area, and can be a
file or a directory.

LP: #1759700

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>