project_loader: stop setting LD_LIBRARY_PATH
Currently, the snapcraft CLI sets LD_LIBRARY_PATH when building classic
snaps so as to ensure libraries within the core snap are found. However,
this wreaks havoc when building classic snaps on Trusty, since it forces
all shelled-out commands to use the libc in the core snap, which causes
segfaults.

Stop setting LD_LIBRARY_PATH, even for classic snaps. Since it's only
set if the host system is determined to be compatible with the libc in
the core snap, this _shouldn't_ have any ill effects.

LP: #1765482

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>