lifecycle: automatically stage dependencies
Take the following situation: we have two parts, `A` and `B`. `A`
depends on `B` by way of having `after: [B]`. Today, if one runs
`snapcraft pull A`, the snapcraft CLI errors out complaining that it
needs to stage `B` first, requiring one to run `snapcraft pull A B`. Not
only is this an error, it's a `RuntimeError`, causing a traceback.

Stop erroring in this situation, and just stage `B` like the user wants.

Fixes SNAPCRAFT-10

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>