travis: use LXD from 3.0 track
An update to LXD was recently released that changes some of the
networking stuff done by `init` such that Travis broke on us. Even
though this breakage also effects 3.0, we should be using LXD 3.0 in CI
so we don't get other random breakage. Also, stop setting up testbr0,
which broke and doesn't seem to be necessary anymore.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>