many: automatically redo step for specified part
Currently, if a step for a given part has already run, the snapcraft CLI
won't run it again without it first being cleaned. For example, if the
`pull` step of part `A` had already run, running `snapcraft pull A`
would say that it had already run, and skip it. Change this behavior so
that, if a part is specified, redo that step (clean and run it again).

LP: #1774260

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>