rust plugin: fix cargo builds and run tests (#2170)
This patch changes it so that we run all instances of cargo
using the build environment. Fixes cases where rustc can't
be found during the pull step.

Also adds 'cargo test' to the build step, to fail the build
if the tests aren't passing. We will only run the tests when we
aren't cross compiling.