project_loader: replace dict keys as well as values (#2166)
The snapcraft CLI actually runs a text replacement over the
`snapcraft.yaml`, replacing all occurrences of its built-in variables
(such as `$SNAPCRAFT_STAGE`) with their actual values, knowing that,
once these are migrated into the final `snap.yaml` the variables have no
meaning. However, when it comes to running the replacements on dicts, it
only checks the values for replacement, not the keys. Fix this by simply
running replacements on the keys as well as the values.

LP: #1746636

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>