plugins: do not install go debs if go snap is used
Right now the go plugin will always pull in the golang-go
build-package. This is problematic when the project also has
a "go" build-snap. The deb package comes before in PATH and it
seems like there is no easy way to override this. This PR fixes
it by simply not installing the golang-go deb if there is a
build-snap already defined.