kbuild: stop modifying kconfigfile
If the `kconfigfile` option is supplied, the kbuild plugin uses a hard
link to put it into place. This becomes problematic if `kconfigfile` is
used alongside `kconfigs`, because doing so modifies the config file,
thereby modifying the original `kconfigfile` in the source tree.

Fix this by copying the `kconfigfile` instead of using hard links.

LP: #1786083

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>