tests: support running spread suite in autopkgtests (#2188)
The spread suite should run as part of the release process. Add an
autopkgtest suite to run it. Also update testing docs for running
autopkgtests on Bionic. Note that these tests don't run on armhf
due to armhf being a LXC container with apparmor disabled,
which breaks snapd. Make the tests isolation-machine so armhf
is skipped.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>