project_loader: add basic template support (#2189)
In a desire to reduce duplication between snaps, introduce support for
templates. Templates have the ability to add YAML snippets to apps, YAML
snippets to parts, and indeed add entire parts to a project. Don't
introduce any actual templates, just the groundwork for them. Only support
them if a base is specified.

LP: #1785109

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>