build providers: environment setup for projects
Add a cloud data generator to the Base Provider which can be
used by any provider to setup the environment properly.

Make the appropriate changes to the project loading logic
to know when it is running in a re-executed mode.

Bring the multipass provider up to speed on the base Provider
changes to make use of this functionality to expose this
functionality for project building.

LP: #1782780
LP: #1782782