plugin API: support command-chain
In order to move from generating wrappers to using command-chain,
plugins need to be able to specify command-chains for their lifecycle
steps as well as dependents and the final snap. While doing this, also
implement functions for obtaining the environment for each lifecycle
step as well as dependents and the final snap.

See https://forum.snapcraft.io/t/6018 for more details.

LP: #1790978

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>