schema: add "legacy" adapter type
Currently, the snapcraft CLI supports app adapters which can either not
be set, or set to "none". If set to "none", no wrappers are generated
for that app. If unset, wrappers are generated.

Change this to default to using a new "legacy" adapter type, which
generates wrappers. This leads nicely to having a "full" adapter for
using the command-chain, which will be coming soon.

LP: #1792018

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>