cli: show trace if no tty
Currently the snapcraft CLI writes any traceback into a tracefile,
regardless of whether or not stdout is a tty. In practice this means
that when running in CI or cleanbuild, errors are hidden within a
tracefile that cannot be accessed.

Check to see if stdout is a tty. If not, print the traceback as well.

LP: #1790697

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>