part grammar processor: lazily capture attributes from plugin
Currently the part processor is handed a plugin and immediately saves
off various grammar-powered attributes, such as stage- and
build-packages, etc. However, there are plugins (specifically Python)
that define these to be properties with definitions that change
depending on the environment. For example, if it already finds Python,
it won't add it to stage-packages. This proves to be problematic when a
part is automatically cleaned because it was dirty: the grammar
processor isn't created multiple times, just once before the part was
cleaned. However, at least in the Python plugin case, once the part is
cleaned it's stage-packages must be re-evaluated or Python will be
missing when it's rebuilt.

Fix this issue by capturing the entire plugin and pulling attributes out
of it when required instead of when the grammar processor is created.

LP: #1794443

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>