plugins: remove the ament plugin when using a base
The ament plugin has been deprecated upstream in favour of a newer
plugin. Use the opportunity of the introduction of bases to remove
support for this plugin from snapcraft.

The spread tests will ensure it continues to work when triggering legacy
mode but will only run a `snapcraft pull` to ensure the plugin is still
reachable. For a full run, the tests should be triggered from the legacy
branch of snapcraft.

LP: #1794775

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>