project_loader: add build-environment part property (#2322)
Beyond using a custom plugin, there is no way for a part to specify
environment variables to be defined during the part's lifecycle. Add a
new property to support doing so declaratively in the snapcraft.yaml,
called `build-environment`. Its usage looks like this:

    parts:
      my-part:
        plugin: <plugin name>
        build-environment:
          - VARIABLE1_NAME: VARIABLE1_VALUE
          - VARIABLE2_NAME: VARIABLE2_VALUE

Note that, as a list, these are exported in order, which allows for
later values to override (or modify) earlier values. Similarly, the
`build-environment` from dependencies will be defined before that of the
dependent's, allowing for it to be modified/overwritten by the
dependent.

LP: #1778174

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>