pluginhandler: library detection instead of injection
With bases, the logic to inject binaries from the host will no longer be
at play. Instead, the base snap for the target project will always be
installed and necessity of the library will be determined throught there
instead

If a library is missing a warning will be issued instead of an error
given the existence of the content snap and its potential use.

LP: #1794556

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>