schema, meta: add "full" app adapter (#2343)
Today, the snapcraft CLI supports both "none" and "legacy" as app
adapters, "none" meaning no wrappers or environment at all, and "legacy"
meaning wrappers. Add a "full" app adapter that means no more wrappers,
but the use of command-chain to accomplish the same thing in a much
nicer way. This requires snapd 2.36 which is not yet released, so
continue using "legacy" as the default for now.

LP: #1794532

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>