cli: handle yaml errors for cleanbuild
When a yaml error occurs, the new code base handles the error as the
lack of a base in the yaml cannot be detected when re-executing.
This strategy has the side-effect of hiding yaml errors behind a
"cleanbuild is not a valid command" scenario with the new cli.
To solve that we add a very thin, hidden, cleanbuild command
to the new cli to work around that situation.

LP: #1803193
Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>