tools: pull in imporved cla_check.py from snapd
We'll get to snap these sometime, but until then, here have the
cla_check.py we're using in snapd, based on yours (but much
improved). It's faster, more accurate, and has nicer, simpler, easier
to understand output.

Note that it does have an output of "I don't know", which still passes
the test. This can happen if a PR is done entirely using github's web
tools from a privacy-enabled account. You'd need to verify manually
then.