repo,baseplugin: support trusting repo keys (#2437)
Backport #2437 to legacy.
LP: #1811304

Update ROS keyring.

Original commit:

Today, snapcraft allows insecure repositories (i.e. repos without a key)
because that is the default behavior in Xenial (core16). This behavior
changed in Zesty, and stays changed in Bionic (core18). Add the ability
to trust keys to the plugin API, take advantage of it in the Catkin
plugin, and flip the switch to only allow snapcraft to use secure
repositories.

LP: #1811304

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>
Signed-off-by: artivis <jeremie.deray@canonical.com>