pluginhandler: handle removal of inconsistent files
When a part is cleaned, the system does its housekeeping by removing
all the files or directories it prepared for stage or prime. If the
user manually removes a file (say, in another part's override-prime),
we should ignore the non-existent file and issue a warning instead
of halting and catching fire.

See https://bugs.launchpad.net/snapcraft/+bug/1813033 for more details.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>