ant, maven and gradle plugins: use correct defaults for jre
Ensure that the expected java version is properly set when the default
openjdk option is used so that the globbing mechanism can format the
path to the java binary correctly.

Also ensure that the sytem works end to end for core16 and core18,
removing the typo of core19 which was supposed to be core16.

LP: #1813637
LP: #1813636
Fixes SNAPCRAFT-D1

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>