cli: retrieve error data from provider
Upon facing an unhandled exception, snapcraft generates traceback
data and stores it in the filesystem. If running in a build provider,
the filesystem location is not accessible from host, and to give user
access to error details we need to retrieve this data from the managed
provider.

This fix addresses this issue by letting inner instance to save error
data in a known location, which is retrieved by outer instance before
proceeding.

LP: #1793082

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>