project loader: remove special LD_LIBRARY_FLAGS handling for classic
This implementation can cause clashes with changes in the host with
regards to the base in use (i.e; core, core18, ...) as their might
be libc6 mismatches at the time that can lead to undefined behavior.

The issue comes up when the core snap is installed as it is when
the libraries are used, this specific behavior is no longer needed
as the library crawling feature has improved since this was first
introduced.

Additionally, the Travis CI job has been moved to using the edge
for the time being to overcome the bootstrapping issue that this
misbehavior creates.

This fix is based on the one introduced in the legacy branch
as snapcraft uses the legacy implementation to build itself.

LP: #1817300

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>