colcon plugin: support build-time chaining
Until recently, colcon had a bug that caused it to not chain workspaces
at build-time. That was recently fixed in a way that broke the colcon
plugin. Rework the plugin to rely more on the setup files and less on
environment variables to avoid this issue in the future.

LP: #1816565

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>