many: support the use of build-aux/snap
Many projects required an alternate snap directory to store
snapcraft related assets, in particular Gnome relate projects
have consolidated with the use of build-aux which snapcraft
is now adopting as an alternative as well.

This also solves the problem where an alternative was needed
in order to build snapd.

LP: #1805219

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>