many: better handling of appstream icons
Appstream icons can be specified as a full path ("local") or just the
application name to be searched in the standard locations ("stock").
The system search for stock icons doesn't include directories that are
part of snap packages, so support stock icons by resolving the path
at packaging time. If icons are not specified in the AppStream file,
use the desktop file definition instead.

LP: #1814898

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>