build providers: improve handling in snap logic
Cleanup up the channel warning when the
SNAPCRAFT_BUILD_ENVIRONMENT_CHANNEL_SNAPCRAFT is not set.
Removed unnecessary use of sudo, given that the providers all run as
root.
Silenced the call on holds for refreshes as it is unnecessarily noisy
and adds little value to users.
Improve the declaration capabilities for providers to have a say on
supporting injection from the host.

LP: #1820864
Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>