Extensions: refactor away gnome exceptions
These are now in project_loader.errors namespace as exceptions that can
be reused. They are also used by the extension loader itself.

Also added support to the extension loader for checking whether the
extension can be used with classic confinement. This check was initially
in the gnome extension itself, but it made sense to pull it out.

Signed-off-by: Daniel Llewellyn <daniel@bowlhat.net>