Appstream: strip ${SNAP} from .desktop icon
A lot of snaps are hacking the desktop file of their projects to insert
a snap-relative path for their icon of the form
`${SNAP}/path/to/icon.png`. The current Appstream implementation does
not understand this format so this commit strips the `${SNAP}` from the
start of the path if it is present before continuing as before.

Signed-off-by: Daniel Llewellyn <daniel@bowlhat.net>