Remove legacy --no-compile option given to pip
The python plugin will now byte compile (create *.pyc files)
when installing packages.

Byte compiling python source in the snap will allow it to start.

The use of --no-commit was marked for removal in the original
commit from more than a year ago so its a good time to remove it.