lifecycle: warn about bases
Add a class that goes throught the project.Project and intersects with the
environment to determine the environment is sane. In the cases where it is
not, a warning will be printed depending on where snapcraft is run
(snap, deb, docker).

The project.ProjectOptions class and private structures have been modified
to provide enhanced messaging for expected acceptable environments.

LP: #1774012
Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>