cli: do not prompt when there is no tty available on stdin
This is a bit of a drive-by. I noticed that inside travis
snapcraft 3 hangs with the following message:
```
Support for 'LXD' needs to be set up. Would you like to do that it now? [y/N]:
```
It seems like we should only prompt here if stdin is attached
to a tty and otherwise just fail early. This should fix hangs
in e.g. scripts.