cli: refactor re-execution into legacy
This refactor makes it easy to still use the old commands regardless of the
existence of a base in the project and to use the store related commands
from non-legacy regardless of where this is run from.

To acheive this, the decision to re-execute into legacy is in the lifecycle
logic by checking for the existence of a base and by providing hidden
commands for the older, previously unavailable, commands.

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>