catkin plugin: check workspace for dependencies
Currently the Catkin plugin doesn't check the workspace for
dependencies, which causes the build to fail if a declared dependency
isn't contained within the rosdep index (even though it's satisfied by
another package in the workspace). Until rosdep grows the ability to do
this on its own, add the ability for the snapcraft CLI to do this
manually by using `rospack` to enumerate the packages contained in the
workspace and filtering dependencies through that list.

LP: #1832044

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>