test_pluginhandler: remove faulty (redundant) tests
Remove tests that previously passed, but were faulty.  They check
chown/chmod calls as proof of ownership/permissions, but the reality
is these files are actually linked, not copied (and updated).  These
mocks were called indirectly from creation of the containing directories,
instead of the desired files.  There are already better tests to validate
the preservation of permissions and ownership, so we'll just rely on those.

Tests removed:
- test_migrate_files_preserves_ownership
- test_migrate_files_chown_permissions

Drop faulty chmod mock assumption
- test_migrate_files_preserves_file_mode_chown_permissions

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>