cli: retrieve error data from provider (#2455)
Adapted from 9d353a6305b8006e0bc0e984c87591e5f08fb09d

Upon facing an unhandled exception, snapcraft generates traceback
data and stores it in the filesystem. If running in a build provider,
the filesystem location is not accessible from host, and to give user
access to error details we need to allow the snapcraft user a chance
to view the traceback before sending the error report.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com
Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>