remote-build: detect and handle early errors on build request
The easiest-to-reproduce example I have come across is to simply
ship the build repository to launchpad without a snapcraft.yaml. In
this case, snapcraft now fails with the following message from launchpad:

```
  Sorry, an error occurred in Snapcraft:
  Remote builder failed with error: 'Cannot find snapcraft.yaml in
  https://cjp256@git.launchpad.net/~cjp256/+git/snapcraft-6cf35d52abc44b19a021b100cb883c8a/ master'
```

The original assumption was that the builder was busy, but now we can
determine if that's truly the case by checking if status is still "Pending".

- A new exception for RemoteBuilderError(err) added.
- Additional sanity checks added for cases that shouldn't occur.
- Added and updated launchpad tests.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>