cli: handle exception when cleaning a part with a fresh project
When the project is fresh (i.e. not-built, or fully cleaned), the
multipass managed instance will fail to find the project and raise a
FileNotFound exception.  Handle it gracefully (ignore it).

- Add new ProjectNotFound exception, set as reportable because it shouldn't
have unhandled cases surfacing up to the user.

- Add test case which would expose the bug if the build provider was multipass.
It's not for unit testing, but we'll add it regardless.

LP #1839014

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>