colcon plugin: add ability to ignore packages
Currently the colcon plugin has the ability to build only a few packages
in a given workspace, or build the entire workspace. However, it's not
unusual to have only a few packages that don't build properly in the
snap. Add the ability to specify packages that should be ignored in the
workspace.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>