remote-build: error if --user is required
When initializing LaunchpadClient, lookup user information from
snapcraft config.  If not present, and user not passed in, raise
error to inform user --user is required. I effectively broke
this behavior in 4253354cf3f995da19f1a568ffcaaa3a1c3359d3 because
this was done in login(), but not consistent with push_source_tree().

- Rename self.user -> self._user for consistency.

- Modify login() and push_source_tree() to use saved username.

- Move launchpad init up before worktree preparation so it fails
  soonest if --user was required.

- Shuffle config handling bits into _load_snapcraft_config() and
  _update_snapcraft_config().

- Update unit tests.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>