multipass: fix setup exception when multipass is not found in PATH
In setup_multipass, there is a case where multipass installs OK, but then
the call to check the version info fails with FileNotFoundError.

The most plausible explanation is that the multipass snap installs OK,
but /snap/bin is not in the PATH, so multipass is not found. This is
easily reproduced.

When this occurs, raise SnapcraftEnvironmentError() with
an error message suggesting the user check their PATH.

Catching this exception was the straw that broke the flake8 "is too
complex" check, so move the wait loop into _wait_for_multipass_ready() -
and flip "multipass" to cls.provider_cmd for consistency.

Fixes SNAPCRAFT-10S

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>