build provider: allow configuration of primary apt mirror
Update cloud-init template to specify the primary mirror.

The template is populated with the (current) default of
`http://archive.ubuntu.com/ubuntu`.

The user may now override this with SNAPCRAFT_APT_PRIMARY_MIRROR.

Only works if used on the command that creates the build VM/conainer,
which executes the cloud-init.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>