project: instantiate snap meta
For now, it just initializes it empty because the yaml configuration
data is mangled by Config and others after this.  Until we converge
to working on snap_meta, we just let Config overwrite the default
instance when the data (mostly) ready. There will be other users
(e.g. snap packaging) that need to converge as well.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>