project: use os.sched_getaffinity instead of multiprocessing.cpu_count
`multiprocessing.cpu_count` always returns the number of raw hardware CPUs
available. However, in the context snapcraft is using it, what is
actually desired is the number of _usable_ CPUs (taking cgroups,
affinity, etc.) into account. For example, if you limit a LXD container
to a single CPU, `multiprocessing.cpu_count` will continue to detect the
number of CPUs available on the host, not 1.

Move snapcraft away from this, and start using `os.sched_getaffinity`
instead, which detects the CPUs actually available for use by the
process.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>